refactor(imports): move app files to dedicated directory
