+++ src/app/app.module.ts
@@ -1,0 +1,8 @@
+import { Module } from "@nestjs/common";
+import { AppController } from "@app/app.controller";
+
+@Module({
+  imports: [],
+  controllers: [AppController]
+})
+export class AppModule {}
